Logica para obtener notas de la base de datos

+++ app/src/main/java/com/gquesada/notes/data/datasources/LocalNoteDataSource.kt
@@ -54,7 +54,7 @@ object LocalNoteDataSource {
         notes.add(localNote)
     }
 
-    fun deleteNote(id: Int) {
+    fun deleteNote(id: Long) {
         notes.removeIf { note -> note.id == id }
     }
