+++ app/src/main/java/com/gquesada/notes/data/datasources/LocalTagDataSource.kt
@@ -21,7 +21,7 @@ object LocalTagDataSource {
         tags.add(tag)
     }
 
-    fun removeTag(id: Int) {
+    fun removeTag(id: Long) {
         tags.removeIf { tag -> tag.id == id }
     }
