+++ app/src/main/java/com/gquesada/notes/data/repositories/TagRepositoryImpl.kt
@@ -18,7 +18,7 @@ class TagRepositoryImpl(
         tagDataSource.addTag(tag.fromModel())
     }
 
-    override fun removeTag(id: Int) {
+    override fun removeTag(id: Long) {
         tagDataSource.removeTag(id)
     }
