+++ app/src/main/java/com/gquesada/notes/domain/usecases/DeleteTagUseCase.kt
@@ -7,7 +7,7 @@ class DeleteTagUseCase(
     private val repository: TagRepository
 ) {
 
-    fun execute(id: Int) {
+    fun execute(id: Long) {
         repository.removeTag(id)
     }
 }
